refactor: rename to use jsx ext and use CamelCase
